+++ Tests/travis/setup-config.sh
@@ -23,7 +23,7 @@ if [ "$INSTALL_TAGSBUNDLE" = "1" ]; then
     # we have to load netgen tags bundle after the Kernel bundles... hopefully OneupFlysystemBundle will stay there :-)
     sed -i 's/OneupFlysystemBundle(),\?/OneupFlysystemBundle(), new Netgen\\TagsBundle\\NetgenTagsBundle(),/' ${APP_DIR}/${EZ_KERNEL}.php
 fi
-# And optionally the EzCoreExtraBundle bundle - not needed any more ?
+# And optionally the EzCoreExtraBundle bundle
 #if grep -q 'lolautruche/ez-core-extra-bundle' composer.lock; then
 if [ "$EZ_VERSION" = "ezplatform2" ]; then
     sed -i 's/OneupFlysystemBundle(),\?/OneupFlysystemBundle(), new Lolautruche\\EzCoreExtraBundle\\EzCoreExtraBundle(),/' ${APP_DIR}/${EZ_KERNEL}.php
@@ -35,9 +35,9 @@ if [ "$EZ_VERSION" = "ezplatform" -o "$EZ_VERSION" = "ezplatform2" ]; then
     sed -i 's/AppBundle(),\?/AppBundle(), new EzSystems\\EzPlatformXmlTextFieldTypeBundle\\EzSystemsEzPlatformXmlTextFieldTypeBundle (),/' ${APP_DIR}/${EZ_KERNEL}.php
 fi
 # Fix the eZ5 autoload configuration for the unexpected directory layout
-sed -i "s#'/../vendor/autoload.php'#'/../../../../vendor/autoload.php'#" ${APP_DIR}/autoload.php
+if [ -f "${APP_DIR}/autoload.php" ]; then sed -i "s#'/../vendor/autoload.php'#'/../../../../vendor/autoload.php'#" ${APP_DIR}/autoload.php fi
 # as well as the config for jms_translation
-sed -i "s#'%kernel.root_dir%/../vendor/ezsystems/ezplatform-admin-ui/src#'%kernel.root_dir%/../../ezplatform-admin-ui/src#" ${APP_DIR}/config.yml
+sed -i "s#'%kernel.root_dir%/../vendor/ezsystems/ezplatform-admin-ui/src#'%kernel.root_dir%/../../ezplatform-admin-ui/src#" ${APP_DIR}/config/config.yml
 
 # Generate legacy autoloads
 if [ "$EZ_VERSION" = "ezpublish-community" ]; then cat Tests/ezpublish-legacy/config.php > vendor/ezsystems/ezpublish-legacy/config.php; fi